修改选课时间，获取选课列表

+++ selclass-server/src/main/java/com/fkczxt/selclassserver/pojo/Cs.java
@@ -27,5 +27,5 @@ public class Cs implements Serializable {
 
     private Integer classid;
 
-    private Integer fraction;
+    private Double fraction;
 }
